requirements.txt
fastapi
uvicorn
sqlmodel
jinja2
python-multipart
psycopg2-binary
python-dotenv